+++ src/test/java/springbook/jdbctemplate/UserServiceTest.java
@@ -11,14 +11,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import springbook.jdbctemplate.service.TxUserService;
+import springbook.jdbctemplate.service.UserServiceImpl;
 
 
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = {DataSourceConfig.class, UserDao.class, UserService.class})
+@ContextConfiguration(classes = {DataSourceConfig.class, UserDao.class, UserServiceImpl.class, TxUserService.class})
 class UserServiceTest {
 
     @Autowired
-    private UserService userService;
+    private TxUserService userService;
 
     @Autowired
     private UserDao userDao;
@@ -29,12 +31,14 @@ class UserServiceTest {
     }
 
     @DisplayName(value = "여러명의 사용자 레벨 정상 업그레이드")
-    void upgradeLevels_succeed() throws Exception {
+    @Test
+    void upgradeLevels_succeed() {
         // given
         User user1 = new User("id1", "name", "password", Level.BASIC);
         User user2 = new User("id2", "name", "password", Level.BASIC);
         User user3 = new User("id3", "name", "password", Level.SILVER);
         List<User> users = List.of(user1, user2, user3);
+        saveAll(users);
 
         // when
         userService.upgradeLevels(users);
@@ -50,12 +54,14 @@ class UserServiceTest {
     }
 
     @DisplayName(value = "하나라도 레벨 정상 업그레이드 되지 않으면 전부 롤백")
-    void upgradeLevels_failed() throws Exception {
+    @Test
+    void upgradeLevels_failed() {
         // given
         User user1 = new User("id1", "name", "password", Level.BASIC);
         User user2 = new User("id2", "name", "password", Level.BASIC);
         User user3 = new User("id3", "name", "password", Level.GOLD);
         List<User> users = List.of(user1, user2, user3);
+        saveAll(users);
 
         // when
         userService.upgradeLevels(users);
@@ -97,5 +103,11 @@ class UserServiceTest {
         // when & then
         assertThatThrownBy(() -> userService.upgradeLevel(user))
                 .isInstanceOf(IllegalStateException.class);
+    }
+
+    private void saveAll(final List<User> users) {
+        for (User user : users) {
+            userDao.save(user);
+        }
     }
 }
